add integration test flag to all integration tests
- rename tests that use a daemon
- move daemon tests to commands_test package
