+++ commands/bootstrap_daemon_test.go
@@ -3,13 +3,15 @@ package commands_test
 import (
 	"testing"
 
-	th "github.com/filecoin-project/go-filecoin/testhelpers"
-
 	"github.com/stretchr/testify/assert"
+
+	th "github.com/filecoin-project/go-filecoin/testhelpers"
+	tf "github.com/filecoin-project/go-filecoin/testhelpers/testflags"
 )
 
 func TestBootstrapList(t *testing.T) {
-	t.Parallel()
+	tf.IntegrationTest(t)
+
 	assert := assert.New(t)
 
 	d := th.NewDaemon(t).Start()
