+++ commands/id_daemon_test.go
@@ -1,16 +1,19 @@
-package commands
+package commands_test
 
 import (
 	"io/ioutil"
 	"testing"
 
-	th "github.com/filecoin-project/go-filecoin/testhelpers"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+
+	th "github.com/filecoin-project/go-filecoin/testhelpers"
+	tf "github.com/filecoin-project/go-filecoin/testhelpers/testflags"
 )
 
 func TestId(t *testing.T) {
-	t.Parallel()
+	tf.IntegrationTest(t)
+
 	assert := assert.New(t)
 
 	d := th.NewDaemon(t).Start()
@@ -25,7 +28,8 @@ func TestId(t *testing.T) {
 }
 
 func TestIdFormat(t *testing.T) {
-	t.Parallel()
+	tf.IntegrationTest(t)
+
 	assert := assert.New(t)
 
 	d := th.NewDaemon(t).Start()
@@ -42,7 +46,8 @@ func TestIdFormat(t *testing.T) {
 }
 
 func TestPersistId(t *testing.T) {
-	t.Parallel()
+	tf.IntegrationTest(t)
+
 	assert := assert.New(t)
 
 	// we need to control this
