+++ commands/mpool_daemon_test.go
@@ -10,10 +10,12 @@ import (
 
 	"github.com/filecoin-project/go-filecoin/fixtures"
 	th "github.com/filecoin-project/go-filecoin/testhelpers"
+	tf "github.com/filecoin-project/go-filecoin/testhelpers/testflags"
 )
 
 func TestMpoolLs(t *testing.T) {
-	t.Parallel()
+	tf.IntegrationTest(t)
+
 	assert := assert.New(t)
 
 	sendMessage := func(d *th.TestDaemon, from string, to string) *th.Output {
@@ -25,7 +27,7 @@ func TestMpoolLs(t *testing.T) {
 	}
 
 	t.Run("return all messages", func(t *testing.T) {
-		t.Parallel()
+
 		d := th.NewDaemon(t, th.KeyFile(fixtures.KeyFilePaths()[0])).Start()
 		defer d.ShutdownSuccess()
 
@@ -50,7 +52,7 @@ func TestMpoolLs(t *testing.T) {
 	})
 
 	t.Run("wait for enough messages", func(t *testing.T) {
-		t.Parallel()
+
 		d := th.NewDaemon(t, th.KeyFile(fixtures.KeyFilePaths()[0])).Start()
 		defer d.ShutdownSuccess()
 
@@ -79,11 +81,12 @@ func TestMpoolLs(t *testing.T) {
 }
 
 func TestMpoolShow(t *testing.T) {
-	t.Parallel()
+	tf.IntegrationTest(t)
+
 	assert := assert.New(t)
 
 	t.Run("shows message", func(t *testing.T) {
-		t.Parallel()
+
 		d := th.NewDaemon(t, th.KeyFile(fixtures.KeyFilePaths()[0])).Start()
 		defer d.ShutdownSuccess()
 
@@ -101,7 +104,7 @@ func TestMpoolShow(t *testing.T) {
 	})
 
 	t.Run("fails missing message", func(t *testing.T) {
-		t.Parallel()
+
 		d := th.NewDaemon(t, th.KeyFile(fixtures.KeyFilePaths()[0])).Start()
 		defer d.ShutdownSuccess()
 
@@ -113,11 +116,12 @@ func TestMpoolShow(t *testing.T) {
 }
 
 func TestMpoolRm(t *testing.T) {
-	t.Parallel()
+	tf.IntegrationTest(t)
+
 	assert := assert.New(t)
 
 	t.Run("remove a message", func(t *testing.T) {
-		t.Parallel()
+
 		d := th.NewDaemon(t, th.KeyFile(fixtures.KeyFilePaths()[0])).Start()
 		defer d.ShutdownSuccess()
