+++ commands/payment_channel_daemon_test.go
@@ -8,21 +8,21 @@ import (
 
 	"github.com/ipfs/go-cid"
 	cbor "github.com/ipfs/go-ipld-cbor"
-
+	"github.com/multiformats/go-multibase"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-
-	"github.com/multiformats/go-multibase"
 
 	"github.com/filecoin-project/go-filecoin/actor/builtin/paymentbroker"
 	"github.com/filecoin-project/go-filecoin/address"
 	"github.com/filecoin-project/go-filecoin/fixtures"
 	th "github.com/filecoin-project/go-filecoin/testhelpers"
+	tf "github.com/filecoin-project/go-filecoin/testhelpers/testflags"
 	"github.com/filecoin-project/go-filecoin/types"
 )
 
 func TestPaymentChannelCreateSuccess(t *testing.T) {
-	t.Parallel()
+	tf.IntegrationTest(t)
+
 	assert := assert.New(t)
 
 	d := makeTestDaemonWithMinerAndStart(t)
@@ -57,12 +57,12 @@ func TestPaymentChannelCreateSuccess(t *testing.T) {
 }
 
 func TestPaymentChannelLs(t *testing.T) {
-	t.Parallel()
+	tf.IntegrationTest(t)
+
 	assert := assert.New(t)
 	require := require.New(t)
 
 	t.Run("Works with default payer", func(t *testing.T) {
-		t.Parallel()
 
 		payer, err := address.NewFromString(fixtures.TestAddresses[2])
 		require.NoError(err)
@@ -80,7 +80,6 @@ func TestPaymentChannelLs(t *testing.T) {
 	})
 
 	t.Run("Works with specified payer", func(t *testing.T) {
-		t.Parallel()
 
 		payer, err := address.NewFromString(fixtures.TestAddresses[2])
 		require.NoError(err)
@@ -102,7 +101,6 @@ func TestPaymentChannelLs(t *testing.T) {
 	})
 
 	t.Run("Notifies when channels not found", func(t *testing.T) {
-		t.Parallel()
 
 		payer, err := address.NewFromString(fixtures.TestAddresses[2])
 		require.NoError(err)
@@ -121,7 +119,8 @@ func TestPaymentChannelLs(t *testing.T) {
 }
 
 func TestPaymentChannelVoucherSuccess(t *testing.T) {
-	t.Parallel()
+	tf.IntegrationTest(t)
+
 	require := require.New(t)
 
 	payer, err := address.NewFromString(fixtures.TestAddresses[2])
@@ -142,7 +141,8 @@ func TestPaymentChannelVoucherSuccess(t *testing.T) {
 }
 
 func TestPaymentChannelRedeemSuccess(t *testing.T) {
-	t.Parallel()
+	tf.IntegrationTest(t)
+
 	require := require.New(t)
 
 	payer, err := address.NewFromString(fixtures.TestAddresses[2])
@@ -177,7 +177,8 @@ func TestPaymentChannelRedeemSuccess(t *testing.T) {
 }
 
 func TestPaymentChannelRedeemTooEarlyFails(t *testing.T) {
-	t.Parallel()
+	tf.IntegrationTest(t)
+
 	require := require.New(t)
 
 	payer, err := address.NewFromString(fixtures.TestAddresses[2])
@@ -213,7 +214,8 @@ func TestPaymentChannelRedeemTooEarlyFails(t *testing.T) {
 }
 
 func TestPaymentChannelReclaimSuccess(t *testing.T) {
-	t.Parallel()
+	tf.IntegrationTest(t)
+
 	require := require.New(t)
 
 	// Initial Balance 10,000
@@ -267,6 +269,8 @@ func TestPaymentChannelReclaimSuccess(t *testing.T) {
 }
 
 func TestPaymentChannelCloseSuccess(t *testing.T) {
+	tf.IntegrationTest(t)
+
 	require := require.New(t)
 
 	// Initial Balance 10,000,000
@@ -317,7 +321,8 @@ func TestPaymentChannelCloseSuccess(t *testing.T) {
 }
 
 func TestPaymentChannelExtendSuccess(t *testing.T) {
-	t.Parallel()
+	tf.IntegrationTest(t)
+
 	require := require.New(t)
 
 	payer, err := address.NewFromString(fixtures.TestAddresses[2])
